Revert to original name and memo definitions (almost)
It turns out that trying to infer ADTs for accounts and transactions
turns isn't worth it. Better to use the API as it's defined. However, it
has become clear that there are cases where merchant and counterparty
are both present and that in those cases it's preferable to use the
merchant. So that's the reason this isn't a total revert - it's also a
refinement.  The presence of both and the need to rank the merchant
higher only became apparent after the exercise of trying to infer an ADT
- so it was still a worthwhile exercise.

Also add --before option.