Remove TODO and commented code for using data protection keychain
Because I now understand there is no way to use it without paying for an Apple
Developer Program membership in order to codesign the binary with an
entitlement such that it has an app ID which can be used as the keychain item's
access group. But the file-based keychain works well enough - it's not worth
the time and money required to benefit from the state being in iCloud.

Even if it had been possible to use the data protection keychain without an
Apple Developer Program membership, it would only have been appropriate to use
it with iCloud sychronisation for the last transactions map anyway - it's not
appropriate to sync access or refresh tokens across devices. They'd have had to
be left unsynchronised in a separate keychain entry.

The benefit of storing state in iCloud would not have been w.r.t.  authz, it
would have been that the last transactions map would be backed up/synchronised
across devices. But that can be solved other ways - ideally by eliminating the
need for it at all...